ModStatR

Motivation: Package compagnon du livre Modélisation statistique par la pratique avec R. Il contient les codes des chapitres du livre ainsi que les solutions des exercices mais aussi d'autres compléments à découvrir.

References

Modélisation statistique par la pratique avec R, Frédéric Bertrand, Emmanuelle Claeys, Myriam Maumy-Bertrand https://www.dunod.com/sciences-techniques/modelisation-statistique-par-pratique-avec-r-cours-et-exercices-corriges, https://github.com/fbertran/ModStatR et https://fbertran.github.io/ModStatR

Examples

set.seed(314)