ModStatR: Statistical Modelling in Action with R

Datasets and functions for the book "Modélisation statistique par la pratique avec R", F. Bertrand, E. Claeys and M. Maumy-Bertrand (2019, ISBN:9782100793525, Dunod, Paris). The first chapter of the book is dedicated to an introduction to the R statistical software. The second chapter deals with correlation analysis: Pearson, Spearman and Kendall simple, multiple and partial correlation coefficients. New wrapper functions for permutation tests or bootstrap of matrices of correlation are provided with the package. The third chapter is dedicated to data exploration with factorial analyses (PCA, CA, MCA, MDA) and clustering. The fourth chapter is dedicated to regression analysis: fitting and model diagnostics are detailed. The exercises focus on covariance analysis, logistic regression, Poisson regression, two-way analysis of variance for fixed or random factors. Various example datasets are shipped with the package: for instance on pokemon, world of warcraft, house tasks or food nutrition analyses.

References

Modélisation statistique par la pratique avec R, Frédéric Bertrand, Emmanuelle Claeys, Myriam Maumy-Bertrand, 2019, ISBN:9782100793525, Dunod, Paris, https://www.dunod.com/sciences-techniques/modelisation-statistique-par-pratique-avec-r-cours-et-exercices-corriges, https://github.com/fbertran/ModStatR/ et https://fbertran.github.io/ModStatR/

See also

Author

Maintainer: Frederic Bertrand frederic.bertrand@utt.fr (ORCID)

Authors:

Emmanuelle Claeys emmanuelle.claeys@unistra.fr

Myriam Maumy-Bertrand myriam.maumy@utt.fr (ORCID)

Examples

set.seed(314)