Intervalles de confiance bootstrap pour une matrice de corrélation de Bravais-Pearson

Intervalles de confiance bootstrap pour une matrice de corrélation de Bravais-Pearson

Usage

boot.mcor.ic(mat, boot.mcor.res, conflevel = 0.95)

Arguments

mat

Matrice des données

boot.mcor.res

Résultat du bootstrap de la matrice de corrélation de Bravais-Pearson

conflevel

Niveau de confiance pour les intervalles

Value

Liste de quatre éléments : matrice des limites inférieures des intervalles de confiance bootstrap percentile, matrice des limites supérieures des intervalles de confiance bootstrap percentile, matrice des limites inférieures des intervalles de confiance bootstrap BCa, matrice des limites supérieures des intervalles de confiance bootstrap BCa

Examples

data(Mesures5,package="BioStatR")
Mes5_red_lr = subset(Mesures5[,-5],subset=Mesures5$espece=="laurier rose")
library(boot)
boot.mcor <- boot(Mes5_red_lr[,c("masse","taille","masse_sec")], rho.mult, R=1000)
boot.mcor
#> 
#> ORDINARY NONPARAMETRIC BOOTSTRAP
#> 
#> 
#> Call:
#> boot(data = Mes5_red_lr[, c("masse", "taille", "masse_sec")], 
#>     statistic = rho.mult, R = 1000)
#> 
#> 
#> Bootstrap Statistics :
#>      original        bias    std. error
#> t1* 0.8407223  0.0002719273  0.02902049
#> t2* 0.6705079 -0.0067273850  0.06751519
#> t3* 0.4150282 -0.0059770037  0.08885902
boot.mcor.ic.res <- boot.mcor.ic(Mes5_red_lr[,c("masse", "taille","masse_sec")],boot.mcor)
boot.mcor.ic.res
#> $cor.ic.percentile.low
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000 0.7748695 0.5073956
#> [2,] 0.7748695 1.0000000 0.2063711
#> [3,] 0.5073956 0.2063711 1.0000000
#> 
#> $cor.ic.percentile.up
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000 0.8944618 0.7771115
#> [2,] 0.8944618 1.0000000 0.5726809
#> [3,] 0.7771115 0.5726809 1.0000000
#> 
#> $cor.ic.BCa.low
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000 0.7721474 0.5072410
#> [2,] 0.7721474 1.0000000 0.2143602
#> [3,] 0.5072410 0.2143602 1.0000000
#> 
#> $cor.ic.BCa.up
#>           [,1]      [,2]      [,3]
#> [1,] 1.0000000 0.8899512 0.7766753
#> [2,] 0.8899512 1.0000000 0.5797118
#> [3,] 0.7766753 0.5797118 1.0000000
#>